allow to provide `namespaces` to the manager CLI
The manager code already has the `watchNamespaces` option and actually
configure it already but there is no way for an user to choose what
namespaces to watch.

For example in Hypershift, CAPO (not deployed by the CAPI operator) runs
in a namespace "capi-provider" and we need ORC to watch for the objects
in that namespace; which without this PR is not possible.

This PR adds `--namespaces` which is a comma-separated list of
namespaces that the controller watches to reconcile ORC objects.
This was taken from CAPO manager code.

(cherry picked from commit 6e0f4a066b7d046b72d08027b98441b12017a0dc)